requirements.txt
chromadb==0.4.22
anthropic==0.18.1
google-generativeai>=0.8.0
pdfplumber==0.10.3
sentence-transformers==2.3.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain==0.1.4
langchain-community==0.0.16
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
Pillow==10.2.0
requests==2.31.0
python-docx==1.1.0
openpyxl==3.1.2
tqdm==4.66.1
pydantic==2.5.3
pydantic-settings==2.1.0
